// csvToJson/types.ts
export type elementType = (string|number)[];

export interface elementProps {
    Site: string
    SiteNumber: number
    Seconds: number
    DateTime: string
    Latitude: number
    Longitude: number
    Hsig: number
    Hmax: number
    Tp: number
    Tz: number
    SST: number
    Direction: number
    Current_Speed: number
    Current_Direction: number
};